use shared library, do not create duplicate permutations, update run.sh

+++ run.sh
@@ -1,11 +1,14 @@
 #!/bin/sh -ex
 
+if [ $# -ne 3 ]; then
+	echo usage: $0 order.xml packlist.xml scoring.xml
+	exit 1
+fi
+
 rm -f score_max
 
-VIEWER=../palletandtruckviewer-3.0/palletViewer
-SCORING=../icra2011TestFiles/scoreAsPlannedConfig1.xml
+python bruteforce2.py $1 | xargs --max-procs=4 python bruteforce3.py $1 $2 $3
 
-python bruteforce2.py $1 \
- | sort\
- | uniq\
- | xargs --max-procs=4 python bruteforce3.py $1 packlist.xml $VIEWER $SCORING
+echo palletViewer -o $1 -p $2 -s $3
+
+echo python evaluate.py $1 $2 $3
